Refactor product listing components to use new sort-filter structure and remove unused utility functions
